pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "outline-tool-backend"
version = "0.1.0"
description = "Outline tool backend"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "BUDA Team"}
]
keywords = ["opensearch", "api", "fastapi", "tibetan"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "opensearch-py>=2.4.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
query = [
    "pyewts>=0.1.1",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/buda-base/outline_tool_backend"
Repository = "https://github.com/buda-base/outline_tool_backend"

[tool.setuptools.packages.find]
where = ["."]
include = ["api*", "scripts*"]
exclude = ["tests*"]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "BLE001",  # Do not catch blind exception
    "C901",    # function too complex
    "COM812",  # Trailing comma missing
    "D",
    "EM101",   # Exception must not use a string literal, assign to variable first
    "EM102",   # Exception must not use an f-string literal, assign to variable first
    "PLR2004", # Magic value used in comparison
    "PLR0911", # Too many return statements
    "PLR0913", # Too many arguments in function definition
    "TC003",   # Move standard library import into a type-checking block
    "TD",      # TODO comments
    "TRY003",  # Avoid specifying long messages outside the exception class    
]